Add placeholder site
Add a more fun place holder "web app" (static site) to the project: a
digital budda board. It's a grid of tiles (divs) when on mouse over
paints the tile with a color from a endless loop of colors. and an
initial transparency value. On subsequent mouse overs, if transparency
present, futher darken the color, other wise proceed as if initial
mouse over

Also add readme for the project. bullet points of considerations that
came up during the project